requirements.txt
Django==2.0.4
Pillow==5.1.0
mysqlclient==1.3.12
django-background-tasks==1.1.13
django-compat==1.0.15
requests==2.18.4
influxdb==5.0.0
gunicorn == 19.7.1
mysqlclient==1.3.12
pyyaml==3.12
aiohttp==2.3.10
SQLAlchemy==1.2.8
PyMySQL==0.8.1
tox==3.0.0
